+++ app/Http/Controllers/Base/LocaleController.php
@@ -3,18 +3,14 @@
 namespace Pterodactyl\Http\Controllers\Base;
 
 use Illuminate\Http\JsonResponse;
-use Illuminate\Translation\Translator;
-use Illuminate\Contracts\Translation\Loader;
 use Pterodactyl\Http\Controllers\Controller;
 use Pterodactyl\Http\Requests\Base\LocaleRequest;
+use Pterodactyl\Services\Translation\FrontendTranslationService;
 
 class LocaleController extends Controller
 {
-    protected Loader $loader;
-
-    public function __construct(Translator $translator)
+    public function __construct(private FrontendTranslationService $translations)
     {
-        $this->loader = $translator->getLoader();
     }
 
     /**
@@ -24,7 +20,7 @@ class LocaleController extends Controller
     {
         $locale = $request->input('locale');
         $namespace = $request->input('namespace');
-        $response[$locale][$namespace] = $this->i18n($this->loader->load($locale, $namespace));
+        $response[$locale][$namespace] = $this->translations->loadNamespace($locale, $namespace);
 
         return new JsonResponse($response, 200, [
             // Cache this in the browser for an hour, and allow the browser to use a stale
@@ -33,32 +29,5 @@ class LocaleController extends Controller
             'Cache-Control' => 'public, max-age=3600, stale-while-revalidate=86400',
             'ETag' => md5(json_encode($response, JSON_THROW_ON_ERROR)),
         ]);
-    }
-
-    /**
-     * Convert standard Laravel translation keys that look like ":foo"
-     * into key structures that are supported by the front-end i18n
-     * library, like "{{foo}}".
-     */
-    protected function i18n(array $data): array
-    {
-        foreach ($data as $key => $value) {
-            if (is_array($value)) {
-                $data[$key] = $this->i18n($value);
-            } else {
-                // Find a Laravel style translation replacement in the string and replace it with
-                // one that the front-end is able to use. This won't always be present, especially
-                // for complex strings or things where we'd never have a backend component anyways.
-                //
-                // For example:
-                // "Hello :name, the :notifications.0.title notification needs :count actions :foo.0.bar."
-                //
-                // Becomes:
-                // "Hello {{name}}, the {{notifications.0.title}} notification needs {{count}} actions {{foo.0.bar}}."
-                $data[$key] = preg_replace('/:([\w.-]+\w)([^\w:]?|$)/m', '{{$1}}$2', $value);
-            }
-        }
-
-        return $data;
     }
 }