+++ resources/views/admin/users/new.blade.php
@@ -51,7 +51,7 @@
                         <div>
                             <select name="language" class="form-control">
                                 @foreach($languages as $key => $value)
-                                    <option value="{{ $key }}" @if(config('app.locale') === $key) selected @endif>{{ $value }}</option>
+                                    <option value="{{ $key }}" @if(old('language', config('app.panel_locale', config('app.locale'))) === $key) selected @endif>{{ $value }}</option>
                                 @endforeach
                             </select>
                             <p class="text-muted"><small>The default language to use when rendering the Panel for this user.</small></p>